Add fontFamily and textColor options
First of all this module is awesome and thanks.

I needed to customize the font family and change the color of the input text when focused, and to do so added `fontFamily` and `textColor` options. (`textColor` maps to `color` in css, I didn't want to accidentally collide with Layer's `color` option).